[flow] move fill_col_dict() to utils module
